requirements.txt
absl-py==0.6.1
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
awscli==1.15.76
backcall==0.1.0
bleach==3.0.2
blinker==1.4
boto3==1.7.75
botocore==1.10.75
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
cryptography-vectors==2.4.2
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.4
django-allauth==0.36.0
django-crispy-forms==1.7.0
django-heroku==0.3.1
django-storages==1.6.6
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
grpcio==1.16.1
gunicorn==19.9.0
h5py==2.8.0
html5lib==1.0.1
hyperlink==18.0.0
icc-rt==2019.0
idna==2.8
incremental==17.5.0
intel-openmp==2019.0
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
numpy==1.15.4
oauthlib==2.1.0
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pydot==1.3.0
Pygments==2.3.1
PyJWT==1.7.1
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.21.0
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
scipy==1.1.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
TBB==0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
Twisted==17.5.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
win-inet-pton==1.0.1
wincertstore==0.2
xlrd==1.2.0
zope.interface==4.6.0
